Renamed 'parser' to 'uncleaner' to better identify it's purpose.
