+++ tests/local/uncleaner/root_uncleaner_test.php
@@ -23,8 +23,8 @@
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-use local_cleanurls\local\urlparser\root_parser;
-use local_cleanurls\local\urlparser\urlparser;
+use local_cleanurls\local\uncleaner\root_uncleaner;
+use local_cleanurls\local\uncleaner\uncleaner;
 
 defined('MOODLE_INTERNAL') || die();
 require_once(__DIR__ . '/../../cleanurls_testcase.php');
@@ -39,21 +39,21 @@ require_once(__DIR__ . '/../../cleanurls_testcase.php');
  */
 class local_cleanurls_urlparser_root_test extends local_cleanurls_testcase {
     public function test_it_exists() {
-        self::assertTrue(class_exists('\local_cleanurls\local\urlparser\root_parser'));
+        self::assertTrue(class_exists('\local_cleanurls\local\uncleaner\root_uncleaner'));
     }
 
     public function test_it_is_a_parser() {
-        $root = new root_parser('/');
-        self::assertInstanceOf(urlparser::class, $root);
+        $root = new root_uncleaner('/');
+        self::assertInstanceOf(uncleaner::class, $root);
     }
 
     public function test_it_takes_a_url_as_string() {
-        $root = new root_parser('/');
+        $root = new root_uncleaner('/');
         self::assertSame('/', $root->get_original_raw_url());
     }
-    
+
     public function test_it_gives_the_clean_url() {
-        $root = new root_parser('/');
+        $root = new root_uncleaner('/');
         $clean = $root->get_clean_url();
         self::assertSame('http://www.example.com/moodle/', $clean->out());
     }
@@ -72,16 +72,16 @@ class local_cleanurls_urlparser_root_test extends local_cleanurls_testcase {
      */
     public function test_it_takes_only_strings($input) {
         $this->expectException(invalid_parameter_exception::class);
-        new root_parser($input);
+        new root_uncleaner($input);
     }
 
     public function test_it_does_not_have_a_parent() {
-        $root = new root_parser('/');
+        $root = new root_uncleaner('/');
         self::assertNull($root->get_parent());
     }
 
     public function test_it_extracts_the_moodle_path() {
-        $root = new root_parser('/abc/def');
+        $root = new root_uncleaner('/abc/def');
         self::assertSame('/moodle', $root->get_moodle_path());
     }
 
@@ -99,7 +99,7 @@ class local_cleanurls_urlparser_root_test extends local_cleanurls_testcase {
      * @dataProvider provider_for_test_it_has_a_subpath
      */
     public function test_it_has_a_subpath($url, $expected) {
-        $root = new root_parser($url);
+        $root = new root_uncleaner($url);
         self::assertSame($expected, $root->get_subpath(), "URL: {$url}");
     }
 
@@ -118,7 +118,7 @@ class local_cleanurls_urlparser_root_test extends local_cleanurls_testcase {
      * @dataProvider provider_for_test_it_has_parameters
      */
     public function test_it_has_parameters($url, $expected) {
-        $root = new root_parser($url);
+        $root = new root_uncleaner($url);
         self::assertSame($expected, $root->get_parameters(), "URL: {$url}");
     }
 
@@ -126,12 +126,12 @@ class local_cleanurls_urlparser_root_test extends local_cleanurls_testcase {
         global $CFG;
 
         $CFG->wwwroot = 'http://moodle.test';
-        $root = new root_parser('/abc/def');
+        $root = new root_uncleaner('/abc/def');
         self::assertSame('', $root->get_moodle_path());
     }
 
     public function test_it_has_its_path_empty() {
-        $root = new root_parser('/hello/world');
+        $root = new root_uncleaner('/hello/world');
         self::assertSame('', $root->get_mypath());
     }
 }
